Refactor app to match the advanced project structure
