+++ dna_web/__init__.py
@@ -1,1 +1,11 @@
-# Archivo __init__.py para el módulo dna_web
+"""Paquete principal de la aplicación DNA Web."""
+
+from dna_web.state import DNAAnalyzer
+from dna_web.pages import index
+from dna_web.components import (
+    navbar,
+    input_section,
+    controls_section,
+    results_section,
+    properties_section,
+)
