+++ dna_web/components/__init__.py
@@ -1,0 +1,7 @@
+"""Archivo __init__.py para el paquete components."""
+
+from dna_web.components.navbar import navbar
+from dna_web.components.input_section import input_section
+from dna_web.components.controls_section import controls_section
+from dna_web.components.results_section import results_section
+from dna_web.components.properties_section import properties_section
