+++ dna_web/pages/__init__.py
@@ -1,0 +1,3 @@
+"""Archivo __init__.py para el paquete pages."""
+
+from dna_web.pages.index import index
